move constants file in webview-ui outside the lib/ directory
