+++ extensions/spectacular/webview-ui/src/constants.ts
